Replace Xaction shorthand with Transaction
