+++ src/main/java/bio/terra/common/PdaoConstant.java
@@ -13,18 +13,18 @@ public final class PdaoConstant {
   public static final String PDAO_INGEST_DATE_COLUMN_ALIAS = PDAO_PREFIX + "ingest_date";
   public static final String PDAO_EXTERNAL_TABLE_PREFIX = PDAO_PREFIX + "etranst_";
   public static final String PDAO_LOAD_HISTORY_TABLE = PDAO_PREFIX + "load_history";
-  public static final String PDAO_XACTIONS_TABLE = PDAO_PREFIX + "transactions";
-  public static final String PDAO_XACTION_ID_COLUMN = PDAO_PREFIX + "transaction_id";
-  public static final String PDAO_XACTION_STATUS_COLUMN = PDAO_PREFIX + "transaction_status";
-  public static final String PDAO_XACTION_DESCRIPTION_COLUMN = PDAO_PREFIX + "transaction_name";
-  public static final String PDAO_XACTION_LOCK_COLUMN = PDAO_PREFIX + "lock";
-  public static final String PDAO_XACTION_CREATED_AT_COLUMN =
+  public static final String PDAO_TRANSACTIONS_TABLE = PDAO_PREFIX + "transactions";
+  public static final String PDAO_TRANSACTION_ID_COLUMN = PDAO_PREFIX + "transaction_id";
+  public static final String PDAO_TRANSACTION_STATUS_COLUMN = PDAO_PREFIX + "transaction_status";
+  public static final String PDAO_TRANSACTION_DESCRIPTION_COLUMN = PDAO_PREFIX + "transaction_name";
+  public static final String PDAO_TRANSACTION_LOCK_COLUMN = PDAO_PREFIX + "lock";
+  public static final String PDAO_TRANSACTION_CREATED_AT_COLUMN =
       PDAO_PREFIX + "transaction_created_at";
-  public static final String PDAO_XACTION_CREATED_BY_COLUMN =
+  public static final String PDAO_TRANSACTION_CREATED_BY_COLUMN =
       PDAO_PREFIX + "transaction_created_by";
-  public static final String PDAO_XACTION_TERMINATED_AT_COLUMN =
+  public static final String PDAO_TRANSACTION_TERMINATED_AT_COLUMN =
       PDAO_PREFIX + "transaction_terminated_at";
-  public static final String PDAO_XACTION_TERMINATED_BY_COLUMN =
+  public static final String PDAO_TRANSACTION_TERMINATED_BY_COLUMN =
       PDAO_PREFIX + "transaction_terminated_by";
   public static final String PDAO_LOAD_HISTORY_STAGING_TABLE_PREFIX =
       PDAO_LOAD_HISTORY_TABLE + "_staging_";
