+++ src/main/java/bio/terra/service/dataset/flight/datadelete/DatasetDataDeleteFlight.java
@@ -10,10 +10,10 @@ import bio.terra.service.configuration.ConfigurationService;
 import bio.terra.service.dataset.DatasetService;
 import bio.terra.service.dataset.flight.LockDatasetStep;
 import bio.terra.service.dataset.flight.UnlockDatasetStep;
-import bio.terra.service.dataset.flight.xactions.TransactionCommitStep;
-import bio.terra.service.dataset.flight.xactions.TransactionLockStep;
-import bio.terra.service.dataset.flight.xactions.TransactionOpenStep;
-import bio.terra.service.dataset.flight.xactions.TransactionUnlockStep;
+import bio.terra.service.dataset.flight.transactions.TransactionCommitStep;
+import bio.terra.service.dataset.flight.transactions.TransactionLockStep;
+import bio.terra.service.dataset.flight.transactions.TransactionOpenStep;
+import bio.terra.service.dataset.flight.transactions.TransactionUnlockStep;
 import bio.terra.service.filedata.flight.ingest.CreateBucketForBigQueryScratchStep;
 import bio.terra.service.filedata.google.gcs.GcsPdao;
 import bio.terra.service.iam.IamAction;
