+++ src/main/java/bio/terra/service/dataset/flight/ingest/DatasetIngestFlight.java
@@ -16,11 +16,11 @@ import bio.terra.service.dataset.DatasetService;
 import bio.terra.service.dataset.DatasetStorageAccountDao;
 import bio.terra.service.dataset.flight.LockDatasetStep;
 import bio.terra.service.dataset.flight.UnlockDatasetStep;
-import bio.terra.service.dataset.flight.xactions.TransactionCommitStep;
-import bio.terra.service.dataset.flight.xactions.TransactionLockStep;
-import bio.terra.service.dataset.flight.xactions.TransactionOpenStep;
-import bio.terra.service.dataset.flight.xactions.TransactionUnlockStep;
-import bio.terra.service.dataset.flight.xactions.upgrade.TransactionUpgradeSingleDatasetStep;
+import bio.terra.service.dataset.flight.transactions.TransactionCommitStep;
+import bio.terra.service.dataset.flight.transactions.TransactionLockStep;
+import bio.terra.service.dataset.flight.transactions.TransactionOpenStep;
+import bio.terra.service.dataset.flight.transactions.TransactionUnlockStep;
+import bio.terra.service.dataset.flight.transactions.upgrade.TransactionUpgradeSingleDatasetStep;
 import bio.terra.service.filedata.FileService;
 import bio.terra.service.filedata.azure.AzureSynapsePdao;
 import bio.terra.service.filedata.azure.blobstore.AzureBlobStorePdao;
