+++ src/main/java/bio/terra/service/dataset/flight/ingest/IngestInsertIntoDatasetTableStep.java
@@ -8,7 +8,7 @@ import bio.terra.model.IngestResponseModel;
 import bio.terra.service.dataset.Dataset;
 import bio.terra.service.dataset.DatasetService;
 import bio.terra.service.dataset.DatasetTable;
-import bio.terra.service.dataset.flight.xactions.TransactionUtils;
+import bio.terra.service.dataset.flight.transactions.TransactionUtils;
 import bio.terra.service.job.JobMapKeys;
 import bio.terra.service.tabulardata.google.BigQueryPdao;
 import bio.terra.stairway.FlightContext;
