+++ src/main/java/bio/terra/service/dataset/flight/transactions/TransactionRollbackFlight.java
@@ -1,9 +1,8 @@
-package bio.terra.service.dataset.flight.xactions;
+package bio.terra.service.dataset.flight.transactions;
 
 import bio.terra.common.CloudPlatformWrapper;
 import bio.terra.common.exception.CommonExceptions;
 import bio.terra.common.iam.AuthenticatedUserRequest;
-import bio.terra.model.TransactionCreateModel;
 import bio.terra.service.dataset.Dataset;
 import bio.terra.service.dataset.DatasetService;
 import bio.terra.service.job.JobMapKeys;
@@ -13,9 +12,9 @@ import bio.terra.stairway.FlightMap;
 import java.util.UUID;
 import org.springframework.context.ApplicationContext;
 
-public class TransactionOpenFlight extends Flight {
+public class TransactionRollbackFlight extends Flight {
 
-  public TransactionOpenFlight(FlightMap inputParameters, Object applicationContext) {
+  public TransactionRollbackFlight(FlightMap inputParameters, Object applicationContext) {
     super(inputParameters, applicationContext);
 
     // get the required daos to pass into the steps
@@ -25,8 +24,8 @@ public class TransactionOpenFlight extends Flight {
 
     UUID datasetId =
         UUID.fromString(inputParameters.get(JobMapKeys.DATASET_ID.getKeyName(), String.class));
-    TransactionCreateModel transactionRequestModel =
-        inputParameters.get(JobMapKeys.REQUEST.getKeyName(), TransactionCreateModel.class);
+    UUID transactionId =
+        UUID.fromString(inputParameters.get(JobMapKeys.TRANSACTION_ID.getKeyName(), String.class));
     Dataset dataset = datasetService.retrieve(datasetId);
     CloudPlatformWrapper cloudPlatform =
         CloudPlatformWrapper.of(dataset.getDatasetSummary().getStorageCloudPlatform());
@@ -34,17 +33,14 @@ public class TransactionOpenFlight extends Flight {
         inputParameters.get(JobMapKeys.AUTH_USER_INFO.getKeyName(), AuthenticatedUserRequest.class);
 
     if (cloudPlatform.isGcp()) {
-      addStep(
-          new TransactionOpenStep(
-              datasetService,
-              bigQueryPdao,
-              userReq,
-              transactionRequestModel.getDescription(),
-              true,
-              true));
+      addStep(new TransactionLockStep(datasetService, bigQueryPdao, transactionId, false, userReq));
+      addStep(new TransactionRollbackMetadataStep(datasetService, bigQueryPdao, transactionId));
+      addStep(new TransactionRollbackDataStep(datasetService, bigQueryPdao, transactionId));
+      addStep(new TransactionRollbackSoftDeleteStep(datasetService, bigQueryPdao, transactionId));
+      addStep(new TransactionRollbackStep(datasetService, bigQueryPdao, transactionId, userReq));
+      addStep(new TransactionUnlockStep(datasetService, bigQueryPdao, transactionId, userReq));
     } else if (cloudPlatform.isAzure()) {
       throw CommonExceptions.TRANSACTIONS_NOT_IMPLEMENTED_IN_AZURE;
     }
-    addStep(new TransactionUnlockStep(datasetService, bigQueryPdao, null, userReq));
   }
 }
