+++ src/main/java/bio/terra/service/dataset/flight/transactions/TransactionRollbackMetadataStep.java
@@ -1,4 +1,4 @@
-package bio.terra.service.dataset.flight.xactions;
+package bio.terra.service.dataset.flight.transactions;
 
 import bio.terra.service.dataset.Dataset;
 import bio.terra.service.dataset.DatasetService;
@@ -12,13 +12,14 @@ import java.util.UUID;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class TransactionRollbackDataStep implements Step {
-  private static final Logger logger = LoggerFactory.getLogger(TransactionRollbackDataStep.class);
+public class TransactionRollbackMetadataStep implements Step {
+  private static final Logger logger =
+      LoggerFactory.getLogger(TransactionRollbackMetadataStep.class);
   private final DatasetService datasetService;
   private final BigQueryPdao bigQueryPdao;
   private final UUID transactionId;
 
-  public TransactionRollbackDataStep(
+  public TransactionRollbackMetadataStep(
       DatasetService datasetService, BigQueryPdao bigQueryPdao, UUID transactionId) {
     this.datasetService = datasetService;
     this.bigQueryPdao = bigQueryPdao;
@@ -29,7 +30,7 @@ public class TransactionRollbackDataStep implements Step {
   public StepResult doStep(FlightContext context) throws InterruptedException {
     Dataset dataset = IngestUtils.getDataset(context, datasetService);
     for (DatasetTable table : dataset.getTables()) {
-      bigQueryPdao.rollbackDatasetTable(dataset, table.getRawTableName(), transactionId);
+      bigQueryPdao.rollbackDatasetMetadataTable(dataset, table, transactionId);
     }
     return StepResult.getStepResultSuccess();
   }
