+++ src/main/java/bio/terra/service/dataset/flight/transactions/TransactionRollbackStep.java
@@ -1,4 +1,4 @@
-package bio.terra.service.dataset.flight.xactions;
+package bio.terra.service.dataset.flight.transactions;
 
 import bio.terra.common.iam.AuthenticatedUserRequest;
 import bio.terra.model.TransactionModel;
@@ -15,43 +15,31 @@ import java.util.UUID;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class TransactionCommitStep implements Step {
-  private static final Logger logger = LoggerFactory.getLogger(TransactionCommitStep.class);
+public class TransactionRollbackStep implements Step {
+  private static final Logger logger = LoggerFactory.getLogger(TransactionRollbackStep.class);
   private final DatasetService datasetService;
   private final BigQueryPdao bigQueryPdao;
-  private final AuthenticatedUserRequest userReq;
-  // If true, saves the transaction as the response to the flight
-  private final boolean returnTransaction;
-  // Can specify the transaction ID at the flight level or through the flight working map using the
-  // "transactionId" key if this value is null
   private final UUID transactionId;
+  private final AuthenticatedUserRequest userReq;
 
-  public TransactionCommitStep(
+  public TransactionRollbackStep(
       DatasetService datasetService,
       BigQueryPdao bigQueryPdao,
-      AuthenticatedUserRequest userReq,
-      boolean returnTransaction,
-      UUID transactionId) {
+      UUID transactionId,
+      AuthenticatedUserRequest userReq) {
     this.datasetService = datasetService;
     this.bigQueryPdao = bigQueryPdao;
-    this.userReq = userReq;
-    this.returnTransaction = returnTransaction;
     this.transactionId = transactionId;
+    this.userReq = userReq;
   }
 
   @Override
   public StepResult doStep(FlightContext context) throws InterruptedException {
     Dataset dataset = IngestUtils.getDataset(context, datasetService);
-    UUID transactionId =
-        this.transactionId == null
-            ? TransactionUtils.getTransactionId(context)
-            : this.transactionId;
     TransactionModel transaction =
         bigQueryPdao.updateTransactionTableStatus(
-            userReq, dataset, transactionId, StatusEnum.COMMITTED);
-    if (returnTransaction) {
-      context.getWorkingMap().put(JobMapKeys.RESPONSE.getKeyName(), transaction);
-    }
+            userReq, dataset, transactionId, StatusEnum.ROLLED_BACK);
+    context.getWorkingMap().put(JobMapKeys.RESPONSE.getKeyName(), transaction);
     return StepResult.getStepResultSuccess();
   }
