+++ src/main/java/bio/terra/service/dataset/flight/transactions/upgrade/TransactionUpgradeSingleDatasetStep.java
@@ -1,9 +1,6 @@
-package bio.terra.service.dataset.flight.xactions;
+package bio.terra.service.dataset.flight.transactions.upgrade;
 
-import bio.terra.service.dataset.Dataset;
 import bio.terra.service.dataset.DatasetService;
-import bio.terra.service.dataset.DatasetTable;
-import bio.terra.service.dataset.flight.ingest.IngestUtils;
 import bio.terra.service.tabulardata.google.BigQueryPdao;
 import bio.terra.stairway.FlightContext;
 import bio.terra.stairway.Step;
@@ -12,26 +9,23 @@ import java.util.UUID;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class TransactionRollbackMetadataStep implements Step {
+public class TransactionUpgradeSingleDatasetStep implements Step {
   private static final Logger logger =
-      LoggerFactory.getLogger(TransactionRollbackMetadataStep.class);
+      LoggerFactory.getLogger(TransactionUpgradeSingleDatasetStep.class);
   private final DatasetService datasetService;
   private final BigQueryPdao bigQueryPdao;
-  private final UUID transactionId;
+  private final UUID datasetId;
 
-  public TransactionRollbackMetadataStep(
-      DatasetService datasetService, BigQueryPdao bigQueryPdao, UUID transactionId) {
+  public TransactionUpgradeSingleDatasetStep(
+      DatasetService datasetService, BigQueryPdao bigQueryPdao, UUID datasetId) {
     this.datasetService = datasetService;
     this.bigQueryPdao = bigQueryPdao;
-    this.transactionId = transactionId;
+    this.datasetId = datasetId;
   }
 
   @Override
   public StepResult doStep(FlightContext context) throws InterruptedException {
-    Dataset dataset = IngestUtils.getDataset(context, datasetService);
-    for (DatasetTable table : dataset.getTables()) {
-      bigQueryPdao.rollbackDatasetMetadataTable(dataset, table, transactionId);
-    }
+    bigQueryPdao.migrateSchemaForTransactions(datasetService.retrieve(datasetId));
     return StepResult.getStepResultSuccess();
   }
